+++ src/stores/shared/index.ts
@@ -1,4 +1,5 @@
 export * from "./collection";
 export * from "./immutable";
 export * from "./merged";
+export * from "./overlay";
 export * from "./schema";