+++ src/stores/shared/schema.ts
@@ -1,11 +1,15 @@
 import tv4, {JsonSchema} from "tv4";
-import {Readable, Writable, writable} from "svelte/store";
+import {Readable, Writable, derived, writable} from "svelte/store";
 
 import {IJSONType} from "../../util/shared/builtin";
-import {IStoreStartStopNotifier, is_writable} from "../../util/shared/stores";
+import {IStoreStartStopNotifier, is_readable, is_writable} from "../../util/shared/stores";
+
+import {overlay} from "./overlay";
 
 /**
  * Returns a `Readable` / `Writable` Svelte Store, that validates values set to the Store, and retrieved via subscriptions, against a JSON Schema
+ *
+ * **NOTE**: If a non-Store value is passed as `store`, then it will be wrapped in a `Writable` Store
  *
  * ```javascript
  * import {schema} from "svelte-commons/lib/stores/shared";
@@ -62,16 +66,16 @@ import {IStoreStartStopNotifier, is_writable} from "../../util/shared/stores";
  * @param start
  */
 export function schema<T extends IJSONType>(
-    store: T | Writable<T>,
+    store: T | Readable<T> | Writable<T>,
     json_schema: JsonSchema,
     start?: IStoreStartStopNotifier<T>
 ): Readable<T> | Writable<T> {
-    if (!is_writable(store)) {
-        store = writable(store, start) as Writable<T>;
-    }
+    // Need to support raw non-Store values being passed, so
+    // we wrap them in a `Writable` Store
+    if (!is_readable(store)) store = writable(store, start);
 
     function _validate(value: T) {
-        if (tv4.validate(value, json_schema)) return true;
+        if (tv4.validate(value, json_schema)) return value;
 
         const {title = "UntitledSchema"} = json_schema;
         const {dataPath, message} = tv4.error;
@@ -79,45 +83,6 @@ export function schema<T extends IJSONType>(
         throw new TypeError(`bad change '${title}${dataPath}' to Schema Store (${message})`);
     }
 
-    if (is_writable(store)) {
-        const {set, subscribe, update} = store as Writable<T>;
-
-        return {
-            set(value) {
-                _validate(value);
-                set(value);
-            },
-
-            subscribe(run, invalidate) {
-                const _run = (value: T) => {
-                    _validate(value);
-                    run(value);
-                };
-
-                return subscribe(_run, invalidate);
-            },
-
-            update(updater) {
-                update((value) => {
-                    value = updater(value);
-
-                    _validate(value);
-                    return value;
-                });
-            }
-        } as Writable<T>;
-    } else {
-        const {subscribe} = store as Readable<T>;
-
-        return {
-            subscribe(run, invalidate) {
-                const _run = (value: T) => {
-                    _validate(value);
-                    run(value);
-                };
-
-                return subscribe(_run, invalidate);
-            }
-        } as Readable<T>;
-    }
+    if (is_writable(store)) return overlay(store, _validate, _validate);
+    else return derived(store, _validate);
 }